requirements.txt
# --- Telegram Bot API ---
aiogram>=3.0.0

# --- Web Server (Webhook) ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# --- Utils ---
python-dotenv>=1.0.0
loguru>=0.7.2
pydantic>=2.6.0